requirements.txt
snowflake-connector-python==3.6.0
pandas==2.1.0
dbt-snowflake==1.7.0